requirements.txt
flask>=3.0
ollama>=0.3
qdrant-client>=1.9
sentence-transformers>=3.0
playwright>=1.40
open-clip-torch>=2.24
torch>=2.0
Pillow>=10.0
